pyproject.toml
[project]
name = "securitize-nav"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv",
    "web3",
]
